requirements.txt
pyyaml
requests
opencv-python
opencv-python-headless
matplotlib
numpy
